fix(KAN-88) P3: move DecideInput interface out of 'use server' file (BUGS-12)
The PR Quality Gate's check-server-action-exports.sh script caught
that actions.ts exported an interface, which Next.js 16+ rejects
at action-invocation time. Moving the interface to a sibling
types.ts module keeps actions.ts pure-async-functions.
